allow sharing of installed packages among users
Created dotspacemacs-use-shared-packages user option, so that if it is
set to non-nil, then directories listed in package-directory-list are
also searched for installed packages in addition to package-user-dir.
Site administrator can then install packages in one or more of the
directories in package-directory-list so that all users at the site can
share the same installed packages without each user having to install
them.

This change has no impact unless dotspacemacs-use-shared-packages is
explicitly changed to non-nil from the default value of nil.