Move SPC p G to SPC p C-g to regenerate tags
I have the impression that SPC p G can be used for something more
useful regarding tags.
Tags regeneration triggers a process that can take a while to make
more sense to make it less accessible.

If there is nothing useful to be added to SPC p G we can consider
reverting this commit.