Use counsel-org-tag when ivy layer is used
When ivy is used, the default org-set-tags will only complete for the
first tag, and any additional tags must be typed manually.  counsel
provides a replacement function to fix this.

Unfortunately, we can't just alias org-set-tags because it is used in
other places, so we must change C-c C-c using a hook.

See https://github.com/syl20bnr/spacemacs/issues/6918