hint about where to put web-mode setq
Web-mode ignores settings set in `dotspacemacs/user-config` on the file opened at launch (though it picks them up for subsequent files), so it took me a while to figure out that I needed to put them in `dotspacemacs/user-init`. This might be nice to save some other people the anguish.