fix broken macOS paths for homebrew service
When Spacemacs is run as a macOS homebrew service, the window-system
at startup time is not set.  This causes the very-early
call to exec-path-from-shell, which was checking for macOS via
window-system, to fail the check.  This is pretty much game over
for macOS and user-customized paths, as the exec paths are not
updated and any extra user-path utilities (e.g. aspell, gls) won't
be found.

This change switches over to using spacemacs/system-is-mac and
spacemacs/system-is-linux as the primary check, and keeps the
window-system check for 'x in case other Unix variants
are getting exec-path set via that check.