Process user-config before command line arguments
emacs-startup-hook is run after the command line arguments are being
processed (and not at all if in batch mode). This has several undesired
side-effects. If Emacs is started with file arguments, user-config is
being run after the respective major mode has been initialized. Thus,
variables set and code defined in user-config will not have the desired
effect in that particular scenario. Also, in batch mode, user-config
will not be evaluated at all which in some cases is not what the user
might want and/or expect.

This refactors the handling of user-config to the after-init-hook and
introduces another needed setup function for init.el.