Fix usage of gorename on symlinked projects
I often work on Go projects under a symlink, e.g. ~/Code/project points to
$GOPATH/src/github.com/grncdr/...

This works fine 99% of the time, but gorename doesn't follow symlinks in order
to get the "truename" of a file, so it will complain about
~/Code/project/somefile.go not belonging to any package. This fixes that
situation by ensuring that gorename is only passed the true filename.