core: Fix which-key second level prefix
`which-key` now interprets `SPC xa` as `SPC` and `xa` instead of `SPC`, `x` and
`a`. This commit inserts additional spaces among the characters in the prefix so
the function `spacemacs/declare-prefix` can work well with `which-key` again.

This fix #7858 .

squash! core: Fix which-key second level prefix

core: Fix major mode prefix