Colors: Major fix to actually render per-theme settings on switch
It isn't enough to just update the variables! We must also refresh the "font
locking" (syntax highlighting) in all buffers that have rainbow-identifiers-mode
currently active, so that they instantly re-paint with their per-theme values.
Otherwise we get stuck with an ugly mishmash of old colors and a new theme.

This change loops through all buffers and marks matching ones for re-painting,
starting with the current buffer first so that the user sees quick results!