Add the "evil-smartparens" package to editing layer
The default implementation of smartparens is not aware of all the various
"extra" editing functions in evil mode. For example, if the caret is at `|` in
the following code:

```
(defun foo (do-thing |bar) (other-thing baz))
```
and a user hits "C" in normal mode, the result is:
```
(defun foo (do-thing |
```
leaving two unbalanced parens. It is much more likely that the user would have
the result be:
```
(defun foo (do-thing |))
```
which is the result following this commit.