Layer common-lisp - set jump handler to SLIME-EDIT-DEFINITION
https://github.com/syl20bnr/spacemacs/issues/8043

"Go to definition" should go to the definition - meaning, where the function is defined in the source code.

I am for changing the default keybinding to `slime-edit-definition` because it fulfills this - allows the user to see the source code for a given function.

`slime-inspect-definition` does not open the source code - it opens the SLIME inspector with the function **object** open in the inspector. While the function **object** contains some valuable data, it does not contain the *definition* itself - which is what the user wants in this case.