select neotree window with backtick key
Frontend
Add ` to the spacemacs root: `,0..9 -> window 0..9.

Hide "` -> select-window-0" in the spacemacs root panel.

Add ` to "[0-9] window N" in the window manipulation transient
state panel.
Change "0-9" to "0..9", and "N" to "0..9" so that it matches the
combined entries in the "SPC" spacemacs root and "SPC b" buffers
panels.
After these changes: [`,0..9] window 0..9

Backend
Bind ` (backtick) to the "select-window-0" command, it selects
an open neotree window.

Add "^^" to the end of the other entries in the first column, so
that the alignment matches the new 3 key entry "[`,0..9]".