improve ledger-reconcile-toggle key binding
Actually, revert @jgertm's change that I asked to make. As @bmag explained, this
is very useful key binding that is similar to C-c C-c as it allows to quickly
access functionality by pressing `dotspacemacs-major-mode-leader-key` twice.
Also note that `dotspacemacs-major-mode-leader-key` can be nil, that's why we
have to handle this value as well.