Fix links in documentation which breaks org html export
Since org 9 internal links must be declared either as #customID
or textTargetOrFullTextSearch. Relative path links must
be either file:relative/path or ./relative/path. Many links
are using the #customID form to trigger a full text search which is not
defined as to the org documentation. In addition relative file paths are
declared as relative/path which is also not defined. If this is done
the html export can fail. Therefore these links must be corrected
for the files where they break the html export.
Fixes #8054 for all spacemacs documentation files.