throw a real error if emacs version is too old
Many people (including me) were confused when trying to install Spacemacs on an
outdated Emacs. A message was issued, but honestly nobody thinks of going to see
the *Messages* buffer when a perfeclty normal vanilla Emacs starts up,
especially not beginners.

This way, an error is visible and users can take corrective actions.