Add sayid, a debugger, to the clojure language
"Sayid (siy EED) is a tool for debugging and profiling clojure code."
(https://bpiel.github.io/sayid/) it allows viewing the results of each line of
code in a clojure function, without editing the code at all.

This commit evilifies the plugin and adds its commands to the "debug" submenu
of the main clojure mode menu.