Clojure refactor: replace cycle coll with convert to coll
Cycle collection type was recently removed from `clj-refactor` as the
feature was migrated and rewritten in `clojure-mode`. The new feature
lets the user convert a collection into a specific collection type
instead of cycling through them.

To added shortcuts correspond to the shortcuts of this feature in
`clojure-mode` and placed in the refactor submenu.