Update bitbucket evil urls to github evil urls
The issues section in the bitbucket evil repository seems to
have been removed when the evil package was migrated to github.

It means that the bitbucket url to the file: evil-states.el,
is eventually going to become outdated and possibly removed.

And the bitbucket issue url now redirects to the bitbucket sign
up page.

Therefore both urls were updated to point to the equivalent
urls in the github evil repository.