Remove CUSTOM ID from documentation to fix #8216
Previously we entered the CUSTOM IDs manually into the
documentation files to fix the html export.
However today this is done automatically during export.
In this case the manually added ids are interfering
with this process and consequently are visible in the
web pages. Therefore I have removed them again.