Update which-key popup: key/command grouping
Comment the code and group the keys/commands based on the
which-key section they appear in.

Reduce the width of the listed keys:
By removing spaces around the two dots `..` in the key sequences, and
by removing the modifier keys before the last number in a sequence.

Update the renamed select-window-[1-9] command name, to the new
winum-select-window-[1-9] name.

Combine keys that call the same command.