New Keybinding for cider-eval-defun-to-comment
Added a Spacemacs style keybinding for the function cider-eval-defun-to-comment

This function evaluates and expression and displays the result as a comment
on the following line.

The CIDER keybinding is `C-C M-;` so the Spacemacs binding uses the `;`
convention, which is also the general character for comments in Emacs.

As this is an evaluation function, the keybinding is placed under the evaluation
part of the major mode menu.

Included documentation in the README.org file