Add `SPC b u` keybinding for reopening killed buffers
This is similar to the `Ctrl+Shift+T` keybinding found in major browsers, and
helps when accidentally killing a buffer (i.e. fat-fingering `SPC b d` when
meaning to press `SPC b s`).

Only buffers that resolve to existing files will be considered, and stored in a
stack which is pushed to and popped from on buffer kill.