Add keybinding to toggle between a keyword & a string
The function `clojure-toggle-keyword-string` will convert a string to a keyword,
or keyword to a string.

The keybinding is added under the Clojure refactor menu, in the
"cycle/clean/convert" section.

`SPC m r c :`

As this is a `clojure-mode` function, it is defined in the clojure-mode
refactoring keybinding section of packages.el.