osx: add option to separate OSX pasteboard from kill-ring
This commit adds a new option `osx-use-separate-clipboards` to the `osx`
layer. When set to non nil, the layer treats the OSX pasteboard separate
from the spacemacs kill-ring. Default is `nil`, which is the same
behaviour as without this commit, hence not expecting any breaking
changes.

This commit also includes changes to the README file, in response to a
previously failing test (document must have a `Features:` section).

Fixes #6533
Fixes #5750