Check if dotspacemacs-directory-snippets-dir exists before adding it
dotspacemacs-directory-snippets-dir contains a path to a directory with yas
snippets in ~/.spacemacs.d (or arbitrary picked dotspacemacs direcotry).
However, while it's not always present (user has to create it manually)
it is always checked while yas is loaded which generates errors.

This diffs forces checking whether the assumed path exists before adding
it to the yas-snippet-dirs.