spacemacs-base: Bind "SPC b l" to buffer-menu-other-window
buffer-menu-other-window is a built-in function which opens a list of
all buffers on the current session of Emacs. With the universal
argument, it opens a list of buffers that are visiting a file.

It works much like "SPC g s": it opens the list on a new window, and can
be dismissed with "q".