Update themes transient state, t for helm-themes
Add the key binding: t
to open helm-themes, without having to use the arrow up key.

Reorder/rename: cycles backward and cycles forward
to: next and previous

Next is listed first, because it's the most common action, and it matches the
order in other transient states.

Change the t argument in the call to spacemacs/cycle-spacemacs-theme, to the
symbol 'backward, to make it clear what the argument does.

Add documentation for the Themes Transient State bindings.